deploy: validações de dados escolares e experiencias re-adicionados com melhorias

+++ src/main/java/com/dbc/vemcv/service/CandidatoCompletoService.java
@@ -58,16 +58,16 @@ public class CandidatoCompletoService {
         CandidatoDTO candidatoCriado = candidatoService.create(candidatoCreateDTO);
         List<DadosEscolaresDTO> dadosEscolaresDTOList = new ArrayList<>();
         List<ExperienciasDTO> experienciasDTOList = new ArrayList<>();
-//        if(candidatoCompletoCreateDTO.getDadosEscolares()!=null&&!candidatoCompletoCreateDTO.getDadosEscolares().isEmpty()){
+        if(candidatoCompletoCreateDTO.getDadosEscolares()!=null&&!candidatoCompletoCreateDTO.getDadosEscolares().isEmpty()){
             for(DadosEscolaresCreateDTO dadosEscolares: candidatoCompletoCreateDTO.getDadosEscolares()){
                 dadosEscolaresDTOList.add(dadosEscolaresService.create(candidatoCriado.getIdCandidato(),dadosEscolares));
             }
-//        }
-//        if(candidatoCompletoCreateDTO.getExperiencias()!=null&&!candidatoCompletoCreateDTO.getExperiencias().isEmpty()) {
+        }
+        if(candidatoCompletoCreateDTO.getExperiencias()!=null&&!candidatoCompletoCreateDTO.getExperiencias().isEmpty()) {
             for (ExperienciasCreateDTO experiencia : candidatoCompletoCreateDTO.getExperiencias()) {
                 experienciasDTOList.add(experienciasService.create(candidatoCriado.getIdCandidato(), experiencia));
             }
-//        }
+        }
         CandidatoCompletoDTO candidatoCompletoDTO = objectMapper.convertValue(candidatoCriado,CandidatoCompletoDTO.class);
         candidatoCompletoDTO.setDadosEscolares(dadosEscolaresDTOList);
         candidatoCompletoDTO.setExperiencias(experienciasDTOList);
@@ -82,16 +82,16 @@ public class CandidatoCompletoService {
         CandidatoDTO candidatoAtualizado = candidatoService.update(idCandidato, candidatoCreateDTO);
         List<DadosEscolaresDTO> dadosEscolaresList = dadosEscolaresService.findByIdCandidato(idCandidato);
         List<ExperienciasDTO> experienciasList = experienciasService.findByIdCandidato(idCandidato);
-//        if(candidatoCompletoCreateDTO.getDadosEscolares()!=null&&!candidatoCompletoCreateDTO.getDadosEscolares().isEmpty()){
+        if(candidatoCompletoCreateDTO.getDadosEscolares()!=null&&!candidatoCompletoCreateDTO.getDadosEscolares().isEmpty()){
             for (DadosEscolaresDTO dadosEscolaresDTO : dadosEscolaresList) {
                 dadosEscolaresService.delete(dadosEscolaresDTO.getIdDadosEscolares());
             }
-//        }
-//        if(candidatoCompletoCreateDTO.getExperiencias()!=null&&!candidatoCompletoCreateDTO.getExperiencias().isEmpty()) {
+        }
+        if(candidatoCompletoCreateDTO.getExperiencias()!=null&&!candidatoCompletoCreateDTO.getExperiencias().isEmpty()) {
             for (ExperienciasDTO experienciasDTO : experienciasList) {
                 experienciasService.delete(experienciasDTO.getIdExperiencia());
             }
-//        }
+        }
 
         List<DadosEscolaresDTO> dadosEscolaresDTOList = new ArrayList<>();
         List<ExperienciasDTO> experienciasDTOList = new ArrayList<>();
